Move APIs out of external folder
