+++ Organizer/__init__.py
@@ -7,7 +7,9 @@ __all__ = [
     "load_info",
     "save_info",
     "Console",
-    "add_info",
+    "add_comicvine_info",
+    "add_metron_info",
+    "add_league_info",
     "Settings",
     "slugify_comic",
     "slugify_publisher",
@@ -17,7 +19,9 @@ __all__ = [
 from Organizer.comic_archive import create_archive, extract_archive
 from Organizer.comic_info import ComicInfo, PublisherInfo, SeriesInfo, load_info, save_info
 from Organizer.console import Console
+from Organizer.settings import Settings
 from Organizer.utils import slugify_comic, slugify_publisher, slugify_series
 
-from Organizer.settings import Settings  # isort:skip
-from Organizer.external import add_info  # isort:skip
+from Organizer.comicvine_api import add_info as add_comicvine_info  # isort: skip
+from Organizer.league_of_comic_geeks_api import add_info as add_league_info  # isort: skip
+from Organizer.metron_api import add_info as add_metron_info  # isort: skip
