+++ Organizer/comicvine_api/__init__.py
@@ -1,0 +1,10 @@
+__all__ = ["add_info", "parse_issue_result", "parse_publisher_result", "parse_volume_result", "Talker"]
+
+from Organizer.comicvine_api.talker import Talker
+
+from Organizer.comicvine_api.comicvine_info import (  # isort: skip
+    add_info,
+    parse_issue_result,
+    parse_publisher_result,
+    parse_volume_result,
+)
