--- Organizer/external/metron_api/__init__.py
@@ -1,9 +1,0 @@
-__all__ = ["add_info", "parse_issue_result", "parse_publisher_result", "parse_series_result", "Talker"]
-
-from Organizer.external.metron_api.metron_talker import Talker  # isort:skip
-from Organizer.external.metron_api.metron_info import (
-    add_info,
-    parse_issue_result,
-    parse_publisher_result,
-    parse_series_result,
-)
