+++ Organizer/league_of_comic_geeks_api/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ["add_info", "parse_comic_result", "Talker"]
+
+from Organizer.league_of_comic_geeks_api.talker import Talker
+
+from Organizer.league_of_comic_geeks_api.league_info import add_info, parse_comic_result  # isort: skip
