+++ Organizer/metron_api/__init__.py
@@ -1,0 +1,10 @@
+__all__ = ["add_info", "parse_issue_result", "parse_publisher_result", "parse_series_result", "Talker"]
+
+from Organizer.metron_api.talker import Talker
+
+from Organizer.metron_api.metron_info import (  # isort: skip
+    add_info,
+    parse_issue_result,
+    parse_publisher_result,
+    parse_series_result,
+)
