+++ tests/test_metron_api.py
@@ -1,7 +1,7 @@
 from datetime import date
 
 from Organizer import ComicInfo, PublisherInfo, SeriesInfo
-from Organizer.external.metron_api import parse_issue_result, parse_publisher_result, parse_series_result
+from Organizer.metron_api import parse_issue_result, parse_publisher_result, parse_series_result
 
 PUBLISHER_ID = 2
 PUBLISHER_TITLE = "DC Comics"
